refactor: move CardType to global const
