+++ src/const.ts
@@ -25,3 +25,9 @@ export const UrlMarker = {
   Default: 'https://assets.htmlacademy.ru/content/intensive/javascript-1/demo/interactive-map/pin.svg',
   Current: 'https://assets.htmlacademy.ru/content/intensive/javascript-1/demo/interactive-map/main-pin.svg'
 };
+
+export enum CardType {
+  Favorites = 'favorites',
+  Cities = 'cities',
+  NearPlaces = 'near-places'
+}
