+++ src/pages/favorites-page/favorites-page.tsx
@@ -1,9 +1,8 @@
 import LocationsItem from '../../components/locations-item/locations-item';
 import Logo from '../../components/logo/logo';
-import { CardType } from '../../components/offer/offer-card/const';
 import OffersList from '../../components/offer/offers-list/offers-list';
 import UserMenu from '../../components/user-menu/user-menu';
-import { CityName } from '../../const';
+import { CardType, CityName } from '../../const';
 import { Offer } from '../../mocks/types/offers';
 
 type FavoritePageProps = {
